# doc/openapi.yaml
openapi: 3.0.0
info:
  title: 医療マスター検索API
  description: 厚生労働省の医療マスター（特定器材、コメント、傷病名、病棟）を検索・取得するためのAPIです。
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: ローカル開発サーバー
paths:
...
  /teeth/search/name:
    get:
      summary: 名称検索
      description: 歯式名称にキーワードを含む歯式を検索します。
      parameters:
        - name: q
          in: query
          description: 検索キーワード（部分一致）
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tooth'
        '400':
          description: パラメータ不正
  /teeth/search/code:
    get:
      summary: 歯式コード検索
      description: 6桁の歯式コードで検索します。
      parameters:
        - name: q
          in: query
          description: 歯式コード（完全一致）
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tooth'
        '400':
          description: パラメータ不正
components:
  schemas:
...
    Tooth:
      type: object
      properties:
        id:
          type: integer
        update_category:
          type: string
          description: "1: 変更区分"
        master_type:
          type: string
          description: "2: マスター種別"
        code:
          type: string
          description: "3: 歯式コード"
        reserved:
          type: string
          description: "4: 予備"
        name:
          type: string
          description: "5: 歯式名称"
        update_date:
          type: string
          description: "6: 変更年月日"
        discontinued_date:
          type: string
          description: "7: 廃止年月日"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time